refactor: promote source download pipelines
