+++ sbir_etl/extractors/source_downloads/__init__.py
@@ -1,0 +1,7 @@
+"""Deterministic source snapshot download pipelines.
+
+Epistemic tier: pipelines. These modules fetch and faithfully materialize
+declared public source files; they do not classify, score, or infer findings.
+"""
+
+EPISTEMIC_TIER = "pipelines"
